daq2/a10gx: Modify interconnect architecture
In case of Nios II based system the main memory interconnect can grow
significantly, because it has to handle all the peripherals and devices
which needs access the memory.

To decrease the size of the interconnect, introduce new secondary avalon
memory mapped bridges between the main interconnect and peripherals.

The following bridges were introduced:

 - avl_mm_peripheral_bridge@0x10050000 - connects all the base design's
   peripherals
 - avl_mm_bridge_tx@0x10060000 - connects all the Avalon interfaces from the TX
   path
 - avl_mm_bridge_rx@0x10068000 - connects all the Avalon interfaces from the RX
   path
 - axi_mm_bridge_0@0x10070000 - connects all the AXI interfaces from TX
   and RX path

Signed-off-by: Dan Hotoleanu <dan.hotoleanu@analog.com>
Signed-off-by: Istvan Csomortani <istvan.csomortani@analog.com>